fix(branch) reconstruct the file
