+++ app/Func/DoublePuyo.js
@@ -3,7 +3,7 @@ class DoublePuyo {
     constructor() {
         this.centerPuyo = new SinglePuyo();
         this.rotatePuyo = new SinglePuyo(this.centerPuyo.col,this.centerPuyo.row-1);
-        this.positions = [true, false, false, false];
+        this.positions = 0;
     }
 }
